updated the `lstm_transducer_stateless` recipes
also revoked previous changes in conformer_ctc3/jit_pretrained.py
