+++ egs/librispeech/ASR/conformer_ctc3/jit_pretrained.py
@@ -24,7 +24,7 @@ Usage (for non-streaming mode):
 (1) ctc-decoding
 ./conformer_ctc3/pretrained.py \
   --nn-model-filename ./conformer_ctc3/exp/cpu_jit.pt \
-  --tokens data/lang_bpe_500/tokens.txt \
+  --bpe-model data/lang_bpe_500/bpe.model \
   --method ctc-decoding \
   --sample-rate 16000 \
   /path/to/foo.wav \
@@ -71,6 +71,7 @@ from typing import List
 
 import k2
 import kaldifeat
+import sentencepiece as spm
 import torch
 import torchaudio
 from decode import get_decoding_params
@@ -115,9 +116,11 @@ def get_parser():
     )
 
     parser.add_argument(
-        "--tokens",
+        "--bpe-model",
         type=str,
-        help="Path to the tokens.txt.",
+        help="""Path to bpe.model.
+        Used only when method is ctc-decoding.
+        """,
     )
 
     parser.add_argument(
@@ -126,9 +129,10 @@ def get_parser():
         default="1best",
         help="""Decoding method.
         Possible values are:
-        (0) ctc-decoding - Use CTC decoding. It uses a tokens.txt file 
-            to convert tokens to actual words or characters. It needs 
-            neither a lexicon nor an n-gram LM.
+        (0) ctc-decoding - Use CTC decoding. It uses a sentence
+            piece model, i.e., lang_dir/bpe.model, to convert
+            word pieces to words. It needs neither a lexicon
+            nor an n-gram LM.
         (1) 1best - Use the best path as decoding output. Only
             the transformer encoder output is used for decoding.
             We call it HLG decoding.
@@ -277,7 +281,6 @@ def main():
     waves = [w.to(device) for w in waves]
 
     logging.info("Decoding started")
-    hyps = []
     features = fbank(waves)
     feature_lengths = [f.size(0) for f in features]
 
@@ -297,16 +300,9 @@ def main():
 
     if params.method == "ctc-decoding":
         logging.info("Use CTC decoding")
+        bpe_model = spm.SentencePieceProcessor()
+        bpe_model.load(params.bpe_model)
         max_token_id = params.num_classes - 1
-
-        # Load tokens.txt here
-        token_table = k2.SymbolTable.from_file(params.tokens)
-
-        def token_ids_to_words(token_ids: List[int]) -> str:
-            text = ""
-            for i in token_ids:
-                text += token_table[i]
-            return text.replace("▁", " ").strip()
 
         H = k2.ctc_topo(
             max_token=max_token_id,
@@ -328,9 +324,9 @@ def main():
         best_path = one_best_decoding(
             lattice=lattice, use_double_scores=params.use_double_scores
         )
-        hyp_tokens = get_texts(best_path)
-        for hyp in hyp_tokens:
-            hyps.append(token_ids_to_words(hyp))
+        token_ids = get_texts(best_path)
+        hyps = bpe_model.decode(token_ids)
+        hyps = [s.split() for s in hyps]
     elif params.method in [
         "1best",
         "nbest-rescoring",
@@ -395,16 +391,16 @@ def main():
             )
             best_path = next(iter(best_path_dict.values()))
 
+        hyps = get_texts(best_path)
         word_sym_table = k2.SymbolTable.from_file(params.words_file)
-        hyp_tokens = get_texts(best_path)
-        for hyp in hyp_tokens:
-            hyps.append(" ".join([word_sym_table[i] for i in hyp]))
+        hyps = [[word_sym_table[i] for i in ids] for ids in hyps]
     else:
         raise ValueError(f"Unsupported decoding method: {params.method}")
 
     s = "\n"
     for filename, hyp in zip(params.sound_files, hyps):
-        s += f"{filename}:\n{hyp}\n\n"
+        words = " ".join(hyp)
+        s += f"{filename}:\n{words}\n\n"
     logging.info(s)
 
     logging.info("Decoding Done")
